Inst: Activate DataTable StateSave
As requested we will now save the state of each Datatable. For this we
will set the option 'StateSave' to true in our global Datatableconfig.

"When enabled a cookie will be used to save table display information
such as pagination information, display length, filtering and sorting.
As such when the end user reloads the page the display will match what
they had previously set up."

[Link]
https://legacy.datatables.net/ref#bStateSave

[Trac]
https://poodle.w-hs.de/trac/ticket/130